Change to parse expressions in style vars
Also added an option to decide whether to enable this feature. The default is enable.
See also to [here](https://github.com/vuejs/rfcs/blob/master/active-rfcs/0043-sfc-style-variables.md).

I tried to handle this feature with a check rule instead of a parser, but it was difficult to get the correct location, so I implement it in the parser.